Rename settings components
