+++ src/app/settings/admin/module.ts
@@ -13,13 +13,13 @@ import {NgModule} from '@angular/core';
 
 import {SharedModule} from '../../shared/shared.module';
 
-import {AdminSettingsComponent} from './admin-settings.component';
-import {AdminSettingsRoutingModule} from './admin-settings.routing.module';
-import {AddAdminDialogComponent} from './admins/add-admin-dialog/add-admin-dialog.component';
-import {AdminsComponent} from './admins/admins.component';
-import {CustomLinksFormComponent} from './custom-link-form/custom-links-form.component';
-import {DatacenterDataDialogComponent} from './dynamic-datacenters/datacenter-data-dialog/datacenter-data-dialog.component';
-import {DynamicDatacentersComponent} from './dynamic-datacenters/dynamic-datacenters.component';
+import {AdminSettingsComponent} from './component';
+import {AdminSettingsRoutingModule} from './routing';
+import {AddAdminDialogComponent} from './admins/add-admin-dialog/component';
+import {AdminsComponent} from './admins/component';
+import {CustomLinksFormComponent} from './custom-link-form/component';
+import {DatacenterDataDialogComponent} from './dynamic-datacenters/datacenter-data-dialog/component';
+import {DynamicDatacentersComponent} from './dynamic-datacenters/component';
 
 @NgModule({
   imports: [SharedModule, AdminSettingsRoutingModule],
